refactor: isCloseTo 확장함수 생성 및 적용
