+++ feature/library/src/main/java/com/into/websoso/feature/library/model/RatingLevelUiModel.kt
@@ -7,12 +7,4 @@ enum class RatingLevelUiModel(
     FOUR(4.0f),
     FOUR_POINT_FIVE(4.5f),
     FOUR_POINT_EIGHT(4.8f),
-    ;
-
-    companion object {
-        internal fun Float.isCloseTo(
-            target: Float,
-            epsilon: Float = 0.01f,
-        ): Boolean = kotlin.math.abs(this - target) < epsilon
-    }
 }
